split the junit package in migration and simplificaiton
