+++ code-transformation/src/main/java/xyz/keksdose/spoon/code_solver/transformations/junit/migration/ExpectedExceptionRemoval.java
@@ -1,5 +1,5 @@
 
-package xyz.keksdose.spoon.code_solver.transformations.junit;
+package xyz.keksdose.spoon.code_solver.transformations.junit.migration;
 
 import java.util.List;
 import java.util.Map.Entry;
@@ -23,6 +23,7 @@ import xyz.keksdose.spoon.code_solver.history.MarkdownString;
 import xyz.keksdose.spoon.code_solver.transformations.BadSmell;
 import xyz.keksdose.spoon.code_solver.transformations.ImportHelper;
 import xyz.keksdose.spoon.code_solver.transformations.TransformationProcessor;
+import xyz.keksdose.spoon.code_solver.transformations.junit.JunitHelper;
 
 public class ExpectedExceptionRemoval extends TransformationProcessor<CtMethod<?>> {
 
@@ -42,6 +43,7 @@ public class ExpectedExceptionRemoval extends TransformationProcessor<CtMethod<?
 			return MarkdownString.fromRaw("ExpectedException");
 		}
 	};
+
 	public ExpectedExceptionRemoval(ChangeListener listener) {
 		super(listener);
 	}
@@ -61,10 +63,10 @@ public class ExpectedExceptionRemoval extends TransformationProcessor<CtMethod<?
 				removeExpectedValue(testAnnotation.get());
 				ImportHelper.addImport("org.junit.jupiter.api.Assertions.assertThrows", true, compilationUnit);
 				setChanged(method.getParent(CtType.class),
-					new Change(EXPECTED_EXCEPTION_BADSMELL,
-						MarkdownString.fromMarkdown(String.format(CHANGE_TEXT_RAW, method.getSimpleName()),
-							String.format(CHANGE_TEXT_MARKDOWN, method.getSimpleName())),
-						method.getParent(CtType.class)));
+						new Change(EXPECTED_EXCEPTION_BADSMELL,
+								MarkdownString.fromMarkdown(String.format(CHANGE_TEXT_RAW, method.getSimpleName()),
+										String.format(CHANGE_TEXT_MARKDOWN, method.getSimpleName())),
+								method.getParent(CtType.class)));
 			}
 		}
 	}
@@ -92,7 +94,7 @@ public class ExpectedExceptionRemoval extends TransformationProcessor<CtMethod<?
 				.createReference("org.junit.jupiter.api.function.Executable");
 		CtExecutableReference<?> assertThrows = getFactory().Executable()
 				.createReference(typeRef, getFactory().Type().voidType(), "assertThrows",
-					List.of(clazzRef, executableJunit));
+						List.of(clazzRef, executableJunit));
 		CtLambda<?> lambda = getFactory().createLambda();
 		lambda.setType((CtTypeReference) getFactory().Type().voidType());
 		lambda.setBody(body);
