+++ code-transformation/src/main/java/xyz/keksdose/spoon/code_solver/transformations/junit/simplification/AssertFalseSameCheck.java
@@ -1,5 +1,5 @@
 
-package xyz.keksdose.spoon.code_solver.transformations.junit;
+package xyz.keksdose.spoon.code_solver.transformations.junit.simplification;
 
 import java.util.List;
 
@@ -16,6 +16,7 @@ import xyz.keksdose.spoon.code_solver.history.Change;
 import xyz.keksdose.spoon.code_solver.history.ChangeListener;
 import xyz.keksdose.spoon.code_solver.transformations.ImportHelper;
 import xyz.keksdose.spoon.code_solver.transformations.TransformationProcessor;
+import xyz.keksdose.spoon.code_solver.transformations.junit.JunitHelper;
 
 public class AssertFalseSameCheck extends TransformationProcessor<CtInvocation<?>> {
 
@@ -32,7 +33,7 @@ public class AssertFalseSameCheck extends TransformationProcessor<CtInvocation<?
 				CtBinaryOperator<?> binaryOperator = (CtBinaryOperator<?>) expression;
 				if (binaryOperator.getKind().equals(BinaryOperatorKind.EQ)) {
 					CtInvocation<?> junit5AssertNotNull = createJunit5AssertSame(binaryOperator.getLeftHandOperand(),
-						binaryOperator.getRightHandOperand());
+							binaryOperator.getRightHandOperand());
 					junit5AssertNotNull.setComments(invocation.getComments());
 					junit5AssertTrue.replace(junit5AssertNotNull);
 					if (invocation.getArguments().size() == 3) {
@@ -76,7 +77,7 @@ public class AssertFalseSameCheck extends TransformationProcessor<CtInvocation<?
 	private void notifyChangeListener(CtInvocation<?> newAssert) {
 		CtType<?> parent = newAssert.getParent(CtType.class);
 		setChanged(parent, new Change(String.format("Replaced assertFalse checking not same with assertNotSame"),
-			"assertFalse with not equals instead of assertNotSame", parent));
+				"assertFalse with not equals instead of assertNotSame", parent));
 	}
 
 }
